+++ src/index.ts
@@ -26,6 +26,7 @@
 
 export * from "./types.js";
 export * from "./constants.js";
+export * from "./env.js";
 export * from "./client.js";
 export * from "./payloads/index.js";
 export * from "./webhook/index.js";